apps/python-services/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
celery[redis]>=5.4
redis>=5.0
asyncpg>=0.29
sqlalchemy[asyncio]>=2.0
pydantic-settings>=2.3
weasyprint>=62.0
Pillow>=10.4
qrcode>=7.4
Jinja2>=3.1
premailer>=3.10
sendgrid>=6.11
twilio>=9.3
pytesseract>=0.3.10
pdfplumber>=0.11
python-magic>=0.4.27
ortools>=9.10
geopy>=2.4
requests>=2.32
lxml>=5.2
scikit-learn>=1.5
xgboost>=2.1
pandas>=2.2
numpy>=1.26
scipy>=1.13
joblib>=1.4
boto3>=1.34
python-multipart>=0.0.9
